+++ lib/src/features/rest_timer/data/rest_timer_data.dart
@@ -1,0 +1,1 @@
+export 'notifications/flutter_local_rest_timer_notification_gateway.dart';